refactor firebase models
