+++ lib/routes/root_page.dart
@@ -1,8 +1,9 @@
+import 'package:firebase_auth_world/model/auth.dart';
 import 'package:flutter/material.dart';
 import 'auth_page.dart';
-import 'package:firebase_auth_world/firebase_helper/auth.dart';
 import 'home_page.dart';
-import 'package:firebase_auth_world/firebase_helper/database.dart';
+import 'package:firebase_auth_world/model/database.dart';
+
 
 class RootPage extends StatefulWidget {
   RootPage({this.auth, this.databse});
